Refactoring Text and Link
* Decoupling `Text`, `Code`, `Del`, `Em`, `Ins`, `Span`, `Strong`, `Sub`
and `Sup` from Ruby and Canvas.
* Decoupling `Link`, `LinkHover`, `LinkUrl` from Ruby and Canvas.
* A bit tricky, `Link` has to be initialized after `Text` class. The
filename is `shoes/types/text_link{.c,.h}` to ensure that.
* Shoes3/shoes3#333
