+++ shoes/app.c
@@ -9,6 +9,8 @@
 #include "shoes/canvas.h"
 #include "shoes/world.h"
 #include "shoes/native/native.h"
+#include "shoes/types/text.h"
+#include "shoes/types/text_link.h"
 
 static void shoes_app_mark(shoes_app *app) {
     shoes_native_slot_mark(app->slot);
