+++ shoes/native/gtk.c
@@ -13,6 +13,8 @@
 #include "shoes/world.h"
 #include "shoes/native/native.h"
 #include "shoes/types/color.h"
+#include "shoes/types/text.h"
+#include "shoes/types/text_link.h"
 #include "shoes/internal.h"
 
 #include <fontconfig/fontconfig.h>
