+++ shoes/native/gtk/gtkswitch.h
@@ -1,8 +1,10 @@
+#include "shoes/ruby.h"
+
 #ifndef SHOES_GTK_SWITCH_H
 #define SHOES_GTK_SWITCH_H
 
 // There is some magical way to create this but let's do this manually for now
-ID s_active;
+SYMBOL_ID(active);
 
 SHOES_CONTROL_REF shoes_native_switch(VALUE self, shoes_canvas *canvas, shoes_place *place, VALUE attr, char *msg);
 void shoes_native_switch_set_active(SHOES_CONTROL_REF ref, gboolean activate);
