+++ shoes/types/types.h
@@ -9,6 +9,8 @@
 #include "shoes/types/spinner.h"
 #include "shoes/types/svg.h"
 #include "shoes/types/switch.h"
+#include "shoes/types/text.h"
+#include "shoes/types/text_link.h"
 #include "shoes/types/timerbase.h"
 #include "shoes/types/video.h"
 
@@ -24,5 +26,7 @@
 	shoes_spinner_init(); \
 	shoes_svg_init(); \
 	shoes_switch_init(); \
+	shoes_text_init(); \
+	shoes_text_link_init(); \
 	shoes_timerbase_init(); \
 	shoes_video_init();
